+++ packages/archive-reader/CHANGELOG.md
@@ -3,6 +3,14 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.0.0](https://github.com/mbret/prose-reader/compare/v13.0.5...v14.0.0) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/archive-reader
+
+
+
+
+
 ## [13.0.0](https://github.com/mbret/prose-reader/compare/v12.0.0...v13.0.0) (2026-09-25)
 
 **Note:** Version bump only for package @prose-reader/archive-reader