+++ packages/react-reader/CHANGELOG.md
@@ -3,6 +3,30 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [14.0.0](https://github.com/mbret/prose-reader/compare/v13.0.5...v14.0.0) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-reader
+
+
+
+
+
+## [13.0.5](https://github.com/mbret/prose-reader/compare/v13.0.4...v13.0.5) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-reader
+
+
+
+
+
+## [13.0.4](https://github.com/mbret/prose-reader/compare/v13.0.3...v13.0.4) (2026-09-25)
+
+**Note:** Version bump only for package @prose-reader/react-reader
+
+
+
+
+
 ## [13.0.3](https://github.com/mbret/prose-reader/compare/v13.0.2...v13.0.3) (2026-09-25)
 
 **Note:** Version bump only for package @prose-reader/react-reader